+++ Source/PLATEAURuntime/Private/RoadNetwork/PLATEAURnDef.cpp
@@ -28,6 +28,16 @@ void FPLATEAURnDef::SetNewObjectWorld(UObject* World)
     NewObjectWorld = World;
 }
 
-inline FVector2D FPLATEAURnDef::To2D(const FVector& Vector) {
+FVector2D FPLATEAURnDef::To2D(const FVector& Vector) {
     return FAxisPlaneEx::ToVector2D(Vector, Plane);
+}
+
+int32 FPLATEAURnDef::Vector3Comparer::operator()(const FVector& A, const FVector& B) const {
+    auto X = Compare(A.X, B.X);
+    if (X != 0)
+        return X;
+    auto Y = Compare(A.Y, B.Y);
+    if (Y != 0)
+        return Y;
+    return X = Compare(A.Z, B.Z);
 }
